git-svn-id: https://svn.code.sf.net/p/imagination/code/trunk@617 6247ab75-7297-438c-a817-b4e4d5db8f37

+++ src/callbacks.c
@@ -1889,7 +1889,7 @@ img_update_stop_display( img_window_struct *img,
 						 gboolean           update_pos )
 {
 	gchar        *string;
-	gint          full;
+	gdouble      full;
 
 	/* Disable/enable slide duration */
 	gtk_widget_set_sensitive( img->duration,
